TIP-428: move readers
