+++ src/Pim/Bundle/BaseConnectorBundle/Reader/Doctrine/ODMProductReader.php
@@ -19,7 +19,7 @@ use Pim\Component\Catalog\Repository\ProductRepositoryInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @deprecated Will be removed in 1.7, please use Pim\Component\Connector\Reader\ProductReader instead.
+ * @deprecated Will be removed in 1.7, please use Pim\Component\Connector\Reader\Database\ProductReader instead.
  */
 class ODMProductReader extends AbstractConfigurableStepElement implements ProductReaderInterface
 {
