+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/readers.yml
@@ -10,7 +10,7 @@ parameters:
     pim_base_connector.reader.cached.class:            Pim\Bundle\BaseConnectorBundle\Reader\CachedReader
 
 services:
-    # Deprecated: Will be removed in 1.7, please use @pim_connector.reader.product instead.
+    # Deprecated: Will be removed in 1.7, please use @pim_connector.reader.database.product instead.
     pim_base_connector.reader.doctrine.product:
         class: '%pim_base_connector.reader.doctrine.product.class%'
         arguments:
