+++ src/Pim/Bundle/ConnectorBundle/Resources/config/readers.yml
@@ -1,7 +1,8 @@
 parameters:
-    pim_connector.reader.product.class:       Pim\Component\Connector\Reader\ProductReader
-    pim_connector.reader.dummy_item.class:    Pim\Component\Connector\Reader\DummyItemReader
-    pim_connector.reader.doctrine.base.class: Pim\Component\Connector\Reader\Doctrine\BaseReader
+    pim_connector.reader.dummy_item.class: Pim\Component\Connector\Reader\DummyItemReader
+
+    pim_connector.reader.database.product.class: Pim\Component\Connector\Reader\Database\ProductReader
+    pim_connector.reader.database.base.class:    Pim\Component\Connector\Reader\Database\BaseReader
 
     pim_connector.reader.file.csv.class:         Pim\Component\Connector\Reader\File\CsvReader
     pim_connector.reader.file.csv_product.class: Pim\Component\Connector\Reader\File\Product\CsvProductReader
@@ -14,8 +15,8 @@ parameters:
     pim_connector.reader.file.media_path_transformer.class: Pim\Component\Connector\Reader\File\Product\MediaPathTransformer
 
 services:
-    pim_connector.reader.product:
-        class: '%pim_connector.reader.product.class%'
+    pim_connector.reader.database.product:
+        class: '%pim_connector.reader.database.product.class%'
         arguments:
             - '@pim_catalog.query.product_query_builder_factory'
             - '@pim_catalog.repository.channel'
